review: move support signals into the pure tally module
