+++ src/django-backend/services/review/tally.py
@@ -7,7 +7,9 @@ matrix. See openspec/changes/less-biased-project-ranking/design.md.
 
 from __future__ import annotations
 
+from collections import defaultdict
 from collections.abc import Collection, Iterable, Sequence
+from dataclasses import dataclass
 from typing import Protocol
 from uuid import UUID
 
@@ -21,6 +23,15 @@ Ballot = Sequence[ProjectId]
 
 # Strength of a non-existent path. Margins are ints, so this never collides.
 _NO_PATH = float("-inf")
+
+
+@dataclass(frozen=True)
+class ProjectSupport:
+    """Raw signals behind one project's placement, for an admin to weigh."""
+
+    first_place_count: int = 0
+    ranked_by_count: int = 0
+    mean_position: float | None = None
 
 
 class OrderingRule(Protocol):
@@ -102,3 +113,38 @@ def _strongest_paths(margins: MarginMatrix) -> dict[ProjectId, dict[ProjectId, f
                 strengths[a][b] = max(strengths[a][b], through)
 
     return strengths
+
+
+def support_signals(
+    ballots: Iterable[Ballot],
+    eligible_project_ids: Collection[ProjectId],
+) -> dict[ProjectId, ProjectSupport]:
+    """First-place count, ranked-by count and mean position among rankers.
+
+    Positions are the reviewer's own contiguous ordering, so a project ranked
+    second on a two-project ballot has position 2 regardless of how many
+    projects the reviewer skipped.
+    """
+    first_place: dict[ProjectId, int] = defaultdict(int)
+    ranked_by: dict[ProjectId, int] = defaultdict(int)
+    position_total: dict[ProjectId, int] = defaultdict(int)
+
+    for ballot in ballots:
+        for position, project_id in enumerate(ballot, start=1):
+            ranked_by[project_id] += 1
+            position_total[project_id] += position
+            if position == 1:
+                first_place[project_id] += 1
+
+    return {
+        project_id: ProjectSupport(
+            first_place_count=first_place[project_id],
+            ranked_by_count=ranked_by[project_id],
+            mean_position=(
+                position_total[project_id] / ranked_by[project_id]
+                if ranked_by[project_id]
+                else None
+            ),
+        )
+        for project_id in eligible_project_ids
+    }
